+++ src/components/TaskListNav/TaskListNav.tsx
@@ -7,7 +7,7 @@ import * as MyModels from 'Store/types';
 import SideBar from 'Components/SideBar';
 import ProjectForm from 'Components/ProjectForm';
 import IProject from 'Entities/project-entities';
-import { fetchProjects } from 'Store/project/project-action-creators';
+import { fetchProjects } from 'Store/project/actions';
 import DropdownCustom from 'Components/DropdownCustom/index';
 
 interface TaskListNavState {
